# SANGAM-PRODUCTION/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SANGAM — Supply Chain Management API",
    "version": "1.0.0",
    "description": "REST API for the SANGAM permissioned blockchain-based supply chain\nmanagement system built for the Indian Army.\n\n## Authentication\nAll protected endpoints require a `Bearer` JWT obtained from `POST /api/auth/login`.\nTokens expire after 8 hours (one operational shift). Use `POST /api/auth/refresh` to rotate.\n\n## Permissions\nEvery endpoint lists the required permission (format: `resource:action`).\nPermissions are granted by role (Day 13 RBAC), temporary delegation (Day 15),\nor single-use emergency override (Day 15).\n\n## Command Scope\nData is filtered to the requesting user's command scope — their unit and all\nsubordinate units. A battalion CO automatically sees all companies and platoons.\nLateral access (sibling units) is denied.\n\n## Offline-First\nAll services degrade gracefully when the database is unavailable.\nNotifications and audit events are buffered in memory and flushed on reconnect.",
    "contact": {
      "name": "SANGAM Development Team",
      "url": "https://github.com/sangam-scm"
    },
    "license": {
      "name": "Proprietary — Indian Army Internal Use"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local / Docker development"
    },
    {
      "url": "https://sangam.army.mil",
      "description": "Production (classified network)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from POST /api/auth/login (Day 14). Expires after 8h."
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "INVALID_REQUEST"
          },
          "message": {
            "type": "string",
            "example": "A descriptive error message"
          }
        }
      },
...
      "RoleName": {
        "type": "string",
        "enum": [
          "SOLDIER",
          "NCO",
          "JCO",
          "LOGISTICS_OFFICER",
          "OFFICER",
          "SENIOR_OFFICER",
          "COMMANDER",
          "AUDITOR",
          "SYSTEM_ADMIN"
        ],
        "description": "Indian Army rank-mapped role (Day 13)"
      },
      "Permission": {
        "type": "string",
        "pattern": "^[a-z]+:[a-z_]+$",
        "example": "supply:approve",
        "description": "resource:action permission string"
      },
      "Role": {
        "type": "object",
        "properties": {
          "name": {
            "$ref": "#/components/schemas/RoleName"
          },
          "displayName": {
            "type": "string"
          },
          "rankLevel": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10
          },
          "description": {
            "type": "string"
          },
          "permissionCount": {
            "type": "integer"
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            }
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid JWT",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions or outside command scope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/rbac/roles": {
      "get": {
        "tags": [
          "RBAC"
        ],
        "summary": "List all army roles",
        "description": "Returns all 9 rank-mapped roles with their permission counts. Requires users:read.",
        "operationId": "listRoles",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Role list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "roles": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Role"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/rbac/my-permissions": {
      "get": {
        "tags": [
          "RBAC"
        ],
        "summary": "Current user permissions",
        "description": "Full permission context for the authenticated user — role info, unit scope, and complete permission list.",
        "operationId": "getMyPermissions",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Permission context",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/rbac/check/{permission}": {
      "get": {
        "tags": [
          "RBAC"
        ],
        "summary": "Permission probe",
        "description": "Check whether the current user holds a specific permission. Used by UIs to conditionally show/hide controls.",
        "operationId": "checkPermission",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "permission",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Permission"
            },
            "example": "supply:approve"
          }
        ],
        "responses": {
          "200": {
            "description": "Permission check result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "permission": {
                      "type": "string"
                    },
                    "granted": {
                      "type": "boolean"
                    },
                    "role": {
                      "$ref": "#/components/schemas/RoleName"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/rbac/audit-logs": {
      "get": {
        "tags": [
          "Audit"
        ],
        "summary": "Query audit logs",
        "description": "Filter and paginate the tamper-evident audit trail. Requires audit:read (SENIOR_OFFICER+).",
        "operationId": "queryAuditLogs",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "action",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "severity",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "INFO",
                "WARNING",
                "CRITICAL",
                "SECURITY"
              ]
            }
          },
          {
            "name": "success",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "startTime",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endTime",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 100,
              "maximum": 1000
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Audit log entries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "entries": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    },
                    "offset": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/rbac/audit-logs/security": {
      "get": {
        "tags": [
          "Audit"
        ],
        "summary": "Security events",
        "description": "Returns SECURITY and CRITICAL severity audit events in the last N hours. Requires audit:read.",
        "operationId": "getSecurityEvents",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "hours",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 24
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Security events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/rbac/audit-logs/verify-integrity": {
      "post": {
        "tags": [
          "Audit"
        ],
        "summary": "Verify audit hash chain",
        "description": "Run an on-demand integrity check of the SHA-256 hash chain. Returns any tampered entries. Requires audit:read.",
        "operationId": "verifyAuditIntegrity",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Integrity result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "verified": {
                      "type": "boolean"
                    },
                    "entriesChecked": {
                      "type": "integer"
                    },
                    "tamperedEntries": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
  },
  "security": []
}